debug

+++ app.py
@@ -181,20 +181,18 @@ async def get_events(host_id: str):
 async def delete_event(host_id: str, event_title: str):
     event_title = urllib.parse.unquote_plus(event_title)
 
-    #response = execute(f'SELECT * FROM punchcard.event WHERE host_id = "{host_id}" AND title = "{event_title}"', 'GET')
-    
-    #if len(response) == 0:
-        #return {'status': 'error', 'message': 'event does not exist'}
     return rds_client.execute_statement(
-            secretArn = params['database_credentials_secret_store_arn'],
-            database = params['database_name'],
-            resourceArn = params['database_cluster_arn'],
-            sql = f'DELETE FROM punchcard.event VALUES(:host_id, :title)',
-            parameters = [
-                {'name': 'host_id', 'value': {'stringValue': host_id}},
-                {'name': 'title', 'value': {'stringValue': event_title}},
-            ]
-        )
+        secretArn = params['database_credentials_secret_store_arn'],
+        database = params['database_name'],
+        resourceArn = params['database_cluster_arn'],
+        sql = f'DELETE FROM punchcard.form',
+    )
+
+    response = execute(f'SELECT * FROM punchcard.event WHERE host_id = "{host_id}" AND title = "{event_title}"', 'GET')
+    
+    if len(response) == 0:
+        return {'status': 'error', 'message': 'event does not exist'}
+
     event_response = execute(f'DELETE FROM punchcard.event WHERE host_id = "{host_id}" AND title = "{event_title}"', 'DELETE')
     form_response = execute(f'DELETE FROM punchcard.form WHERE host_id = "{host_id}" AND event_title = "{event_title}"', 'DELETE')
